Move some packages around
