+++ build.gradle
@@ -120,7 +120,7 @@ license {
 }
 
 sortClassFields {
-    add 'main', 'com.almuradev.almura.content.block.sound.BlockSoundGroups'
+    add 'main', 'com.almuradev.almura.content.block.component.sound.BlockSoundGroups'
     add 'main', 'com.almuradev.almura.content.item.group.ItemGroups'
     add 'main', 'com.almuradev.almura.content.material.MapColors'
     add 'main', 'com.almuradev.almura.content.material.Materials'
