+++ src/main/java/com/almuradev/almura/CommonProxy.java
@@ -12,12 +12,12 @@ import com.almuradev.almura.content.AssetType;
 import com.almuradev.almura.content.block.BuildableBlockType;
 import com.almuradev.almura.content.block.builder.AbstractBlockTypeBuilder;
 import com.almuradev.almura.content.block.builder.rotatable.HorizontalTypeBuilderImpl;
-import com.almuradev.almura.content.block.data.blockbreak.BlockBreak;
-import com.almuradev.almura.content.block.data.blockbreak.drop.Drop;
-import com.almuradev.almura.content.block.data.blockbreak.drop.ExperienceDrop;
+import com.almuradev.almura.content.block.component.action.blockbreak.BlockBreak;
+import com.almuradev.almura.content.block.component.action.blockbreak.drop.Drop;
+import com.almuradev.almura.content.block.component.action.blockbreak.drop.ExperienceDrop;
+import com.almuradev.almura.content.block.component.sound.BlockSoundGroup;
+import com.almuradev.almura.content.block.component.sound.BlockSoundGroupBuilder;
 import com.almuradev.almura.content.block.rotatable.HorizontalType;
-import com.almuradev.almura.content.block.sound.BlockSoundGroup;
-import com.almuradev.almura.content.block.sound.BlockSoundGroupBuilder;
 import com.almuradev.almura.content.item.BuildableItemType;
 import com.almuradev.almura.content.item.builder.AbstractItemTypeBuilder;
 import com.almuradev.almura.content.item.group.ItemGroup;
