+++ src/main/java/com/almuradev/almura/asm/mixin/core/block/MixinAlmuraBlock.java
@@ -9,13 +9,13 @@ import com.almuradev.almura.asm.mixin.interfaces.IMixinAlmuraBlock;
 import com.almuradev.almura.asm.mixin.interfaces.IMixinDelegateBlockAttributes;
 import com.almuradev.almura.asm.mixin.interfaces.IMixinDelegateMaterialAttributes;
 import com.almuradev.almura.content.block.BuildableBlockType;
-import com.almuradev.almura.content.block.data.Action;
-import com.almuradev.almura.content.block.data.blockbreak.BlockBreak;
-import com.almuradev.almura.content.block.data.blockbreak.drop.Drop;
-import com.almuradev.almura.content.block.data.blockbreak.drop.ItemDrop;
+import com.almuradev.almura.content.block.component.action.Action;
+import com.almuradev.almura.content.block.component.action.blockbreak.BlockBreak;
+import com.almuradev.almura.content.block.component.action.blockbreak.drop.Drop;
+import com.almuradev.almura.content.block.component.action.blockbreak.drop.ItemDrop;
+import com.almuradev.almura.content.block.component.sound.BlockSoundGroup;
 import com.almuradev.almura.content.block.impl.GenericBlock;
 import com.almuradev.almura.content.block.impl.rotatable.GenericHorizontalBlock;
-import com.almuradev.almura.content.block.sound.BlockSoundGroup;
 import com.almuradev.almura.content.item.group.ItemGroup;
 import com.almuradev.almura.content.loader.CatalogDelegate;
 import net.minecraft.block.Block;
