+++ src/main/java/com/almuradev/almura/asm/mixin/core/block/MixinBlock.java
@@ -8,8 +8,8 @@ package com.almuradev.almura.asm.mixin.core.block;
 import com.almuradev.almura.content.block.BuildableBlockType;
 import com.almuradev.almura.content.item.group.ItemGroup;
 import net.minecraft.block.Block;
-import net.minecraft.block.state.IBlockState;
 import net.minecraft.block.SoundType;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
