+++ src/main/java/com/almuradev/almura/asm/mixin/interfaces/IMixinAlmuraBlock.java
@@ -5,7 +5,7 @@
  */
 package com.almuradev.almura.asm.mixin.interfaces;
 
-import com.almuradev.almura.content.block.data.blockbreak.BlockBreak;
+import com.almuradev.almura.content.block.component.action.blockbreak.BlockBreak;
 
 import java.util.List;
