+++ src/main/java/com/almuradev/almura/asm/mixin/interfaces/IMixinDelegateBlockAttributes.java
@@ -5,7 +5,7 @@
  */
 package com.almuradev.almura.asm.mixin.interfaces;
 
-import com.almuradev.almura.content.block.sound.BlockSoundGroup;
+import com.almuradev.almura.content.block.component.sound.BlockSoundGroup;
 import com.almuradev.almura.content.loader.CatalogDelegate;
 
 public interface IMixinDelegateBlockAttributes {
