+++ src/main/java/com/almuradev/almura/configuration/MappedConfigurationAdapter.java
@@ -14,7 +14,6 @@ import ninja.leaping.configurate.objectmapping.ObjectMapper;
 import ninja.leaping.configurate.objectmapping.ObjectMappingException;
 
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 
 public final class MappedConfigurationAdapter<T extends AbstractConfiguration> {
