+++ src/main/java/com/almuradev/almura/content/AssetType.java
@@ -6,8 +6,8 @@
 package com.almuradev.almura.content;
 
 import com.almuradev.almura.content.block.BuildableBlockType;
+import com.almuradev.almura.content.block.component.sound.BlockSoundGroup;
 import com.almuradev.almura.content.block.rotatable.HorizontalType;
-import com.almuradev.almura.content.block.sound.BlockSoundGroup;
 import com.almuradev.almura.content.item.BuildableItemType;
 import com.almuradev.almura.content.item.group.ItemGroup;
 import com.almuradev.almura.registry.BuildableCatalogType;
