+++ src/main/java/com/almuradev/almura/content/block/component/action/blockbreak/action/package-info.java
@@ -4,4 +4,4 @@
  * Copyright (c) AlmuraDev <http://github.com/AlmuraDev/>
  */
 @org.spongepowered.api.util.annotation.NonnullByDefault
-package com.almuradev.almura.content.block.data.blockbreak.drop;
+package com.almuradev.almura.content.block.component.action.blockbreak.action;
