+++ src/main/java/com/almuradev/almura/content/block/component/action/blockbreak/drop/ItemDrop.java
@@ -3,7 +3,7 @@
  *
  * Copyright (c) AlmuraDev <http://github.com/AlmuraDev/>
  */
-package com.almuradev.almura.content.block.data.blockbreak.drop;
+package com.almuradev.almura.content.block.component.action.blockbreak.drop;
 
 import com.almuradev.almura.content.loader.CatalogDelegate;
 import com.google.common.base.MoreObjects;
