+++ src/main/java/com/almuradev/almura/content/block/component/sound/BlockSoundGroupBuilder.java
@@ -3,7 +3,7 @@
  *
  * Copyright (c) AlmuraDev <http://github.com/AlmuraDev/>
  */
-package com.almuradev.almura.content.block.sound;
+package com.almuradev.almura.content.block.component.sound;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
