+++ src/main/java/com/almuradev/almura/content/loader/task/SetBlockAttributesTask.java
@@ -8,10 +8,9 @@ package com.almuradev.almura.content.loader.task;
 import com.almuradev.almura.Almura;
 import com.almuradev.almura.Constants;
 import com.almuradev.almura.content.block.BuildableBlockType;
+import com.almuradev.almura.content.block.component.action.blockbreak.BlockBreakSerializer;
+import com.almuradev.almura.content.block.component.sound.BlockSoundGroup;
 import com.almuradev.almura.content.block.impl.BlockAABB;
-import com.almuradev.almura.content.Pack;
-import com.almuradev.almura.content.block.data.blockbreak.BlockBreakSerializer;
-import com.almuradev.almura.content.block.sound.BlockSoundGroup;
 import com.almuradev.almura.content.loader.Asset;
 import com.almuradev.almura.content.loader.AssetContext;
 import com.almuradev.almura.content.loader.CatalogDelegate;
