+++ src/main/resources/META-INF/almura_at.cfg
@@ -1,4 +1,4 @@
-public-f com.almuradev.almura.content.block.sound.BlockSoundGroups *
+public-f com.almuradev.almura.content.block.component.sound.BlockSoundGroups *
 public-f com.almuradev.almura.content.item.group.ItemGroups *
 public-f com.almuradev.almura.content.material.MapColors *
 public-f com.almuradev.almura.content.material.Materials *
